Don't call defaults command when not on OSX

+++ .bashrc
@@ -92,9 +92,10 @@ export FINDBUGS_HOME=/usr/local/Cellar/findbugs/3.0.1/libexec
 # Have this last to make sure my own bins are searched first
 # PATH="$HOME/bin:$HOME/bin/git-radar:$PATH"
 
-
-# Decrease animation duration when using expose
-defaults write com.apple.dock expose-animation-duration -float 0.2; killall Dock
+if [ "$(uname -s)" == "Darwin" ] ; then
+    # Decrease animation duration when using expose
+    defaults write com.apple.dock expose-animation-duration -float 0.2; killall Dock
+fi
 
 #THIS MUST BE AT THE END OF THE FILE FOR GVM TO WORK!!!
 export GVM_DIR="/Users/nicolai/.gvm"
